+++ Projects/HpAdaptivity/Tools.cpp
@@ -410,7 +410,7 @@ TPZGeoMesh * MalhaGeoT(const int h,bool hrefine){//malha triangulo
 		const int dim = 2;//AQUI
 		
 		REAL co[nnode][dim] ={{0.,0.},{1.,0.},{1.,1.},{0.,1.}};// {{0.,0.},{2.,0},{2.,2.},{0.,2.}};//{{-2.,-2},{2.,-2},{2.,2.},{-2.,2.}};//
-		int indices[2][nnode];//como serao enumerados os nos
+		long indices[2][nnode];//como serao enumerados os nos
 		
 		//el 1
 		indices[0][0] = 0;
@@ -425,7 +425,7 @@ TPZGeoMesh * MalhaGeoT(const int h,bool hrefine){//malha triangulo
 		int nod;
 		TPZVec<REAL> coord(dim);
 		for(nod=0; nod<nnode; nod++) {
-				int nodind = gmesh->NodeVec().AllocateNewElement();
+				long nodind = gmesh->NodeVec().AllocateNewElement();
 				
 				for(int d = 0; d < dim; d++)
 				{
@@ -436,14 +436,14 @@ TPZGeoMesh * MalhaGeoT(const int h,bool hrefine){//malha triangulo
 		//Criacao de elementos
 		
 		
-		TPZVec<int> nodind1(3);
-		TPZVec<int> nodind2(3);
+		TPZVec<long> nodind1(3);
+		TPZVec<long> nodind2(3);
 		for(int i=0; i<3; i++){
 				nodind1[i] = indices[0][i];
 				nodind2[i] = indices[1][i];
 		}
 		
-		int index;
+		long index;
 		elvec[0] = gmesh->CreateGeoElement(ETriangle,nodind1,1,index); //AQUI
 		elvec[1] = gmesh->CreateGeoElement(ETriangle,nodind2,1,index); //AQUI
 		
@@ -466,8 +466,8 @@ TPZGeoMesh * MalhaGeoT(const int h,bool hrefine){//malha triangulo
 		//	Refinamento uniforme
 		for(int ref = 0; ref < h; ref++){// h indica o numero de refinamentos
 				TPZVec<TPZGeoEl *> filhos;
-				int n = gmesh->NElements();
-				for(int i = 0; i < n; i++){
+				long n = gmesh->NElements();
+				for(long i = 0; i < n; i++){
 						TPZGeoEl * gel = gmesh->ElementVec()[i];
 						if(!gel->HasSubElement())
 						{
@@ -480,10 +480,10 @@ TPZGeoMesh * MalhaGeoT(const int h,bool hrefine){//malha triangulo
 				{
 						
 						TPZVec<TPZGeoEl *> filhos;
-						int n = gmesh->NElements();
+						long n = gmesh->NElements();
 						
 						
-						for(int i = 0; i < n; i++){	
+						for(long i = 0; i < n; i++){	
 								TPZGeoEl * gel = gmesh->ElementVec()[i];
 								if(!gel->HasSubElement() && gel->Dimension()==2 && i%2==0)
 										
@@ -498,10 +498,10 @@ TPZGeoMesh * MalhaGeoT(const int h,bool hrefine){//malha triangulo
 				{
 						
 						TPZVec<TPZGeoEl *> filhos;
-						int n = gmesh->NElements();
+						long n = gmesh->NElements();
 						
 						
-						for(int i = 0; i < n; i++){	
+						for(long i = 0; i < n; i++){	
 								TPZGeoEl * gel = gmesh->ElementVec()[i];
 								if (gel->Dimension()!=1) {
 										continue;
@@ -540,7 +540,7 @@ TPZGeoMesh * MalhaGeo/*QUADRILATEROS*/ ( const int h, bool hrefine)
 		int nod;
 		for ( nod=0; nod<nnode; nod++ )
 		{
-				int nodind = gmesh->NodeVec().AllocateNewElement();
+				long nodind = gmesh->NodeVec().AllocateNewElement();
 				TPZVec<REAL> coord ( 2 );
 				coord[0] = co[nod][0];
 				coord[1] = co[nod][1];
@@ -550,9 +550,9 @@ TPZGeoMesh * MalhaGeo/*QUADRILATEROS*/ ( const int h, bool hrefine)
 		int el;
 		for ( el=0; el<nelem; el++ )
 		{
-				TPZVec<int> nodind ( 4 );
+				TPZVec<long> nodind ( 4 );
 				for ( nod=0; nod<4; nod++ ) nodind[nod]=indices[el][nod];
-				int index;
+				long index;
 				elvec[el] = gmesh->CreateGeoElement ( EQuadrilateral,nodind,1,index );
 		}
 		
@@ -567,8 +567,8 @@ TPZGeoMesh * MalhaGeo/*QUADRILATEROS*/ ( const int h, bool hrefine)
 		for ( int ref = 0; ref < h; ref++ )
 		{// h indica o numero de refinamentos
 				TPZVec<TPZGeoEl *> filhos;
-				int n = gmesh->NElements();
-				for ( int i = 0; i < n; i++ )
+				long n = gmesh->NElements();
+				for ( long i = 0; i < n; i++ )
 				{
 						TPZGeoEl * gel = gmesh->ElementVec() [i];
 						//if ( gel->Dimension() == 2 ) gel->Divide ( filhos );
@@ -584,10 +584,10 @@ TPZGeoMesh * MalhaGeo/*QUADRILATEROS*/ ( const int h, bool hrefine)
 				{
 						
 						TPZVec<TPZGeoEl *> filhos;
-						int n = gmesh->NElements();
+						long n = gmesh->NElements();
 						
 						
-						for(int i = 0; i < n; i++){	
+						for(long i = 0; i < n; i++){	
 								TPZGeoEl * gel = gmesh->ElementVec()[i];
 								if(!gel->HasSubElement() && gel->Dimension()==2 && i%2==0)
 										
@@ -602,10 +602,10 @@ TPZGeoMesh * MalhaGeo/*QUADRILATEROS*/ ( const int h, bool hrefine)
 				{
 						
 						TPZVec<TPZGeoEl *> filhos;
-						int n = gmesh->NElements();
+						long n = gmesh->NElements();
 						
 						
-						for(int i = 0; i < n; i++){	
+						for(long i = 0; i < n; i++){	
 								TPZGeoEl * gel = gmesh->ElementVec()[i];
 								if (gel->Dimension()!=1) {
 										continue;
@@ -651,7 +651,7 @@ TPZGeoMesh * MalhaGeo2(const int h){//malha quadrilatero com 4 elementos
 		int nod;
 		TPZVec<REAL> coord(dim);
 		for(nod=0; nod<nnode; nod++) {
-				int nodind = gmesh->NodeVec().AllocateNewElement();
+				long nodind = gmesh->NodeVec().AllocateNewElement();
 				
 				for(int d = 0; d < dim; d++)
 				{
@@ -661,9 +661,9 @@ TPZGeoMesh * MalhaGeo2(const int h){//malha quadrilatero com 4 elementos
 		}
 		//Criacao de elementos
 		int matId=40;
-		int id=0;
-		TPZVec <int> TopolQuad(4);
-		TPZVec <int> TopolLine(2);
+		long id=0;
+		TPZVec <long> TopolQuad(4);
+		TPZVec <long> TopolLine(2);
 		//-----
 		
         TopolQuad[0] = 0;
@@ -798,7 +798,7 @@ TPZGeoMesh * MalhaGeo2(const int h){//malha quadrilatero com 4 elementos
 		
 }
 void SolGraf(TPZCompMesh *malha, std::ofstream &GraficoSol){
-		const int nelem = malha->NElements();
+		const long nelem = malha->NElements();
 		
 		//TPZFMatrix<REAL> sol(4,1,0.);
 		TPZManVector<REAL,4> solP(1);
@@ -816,7 +816,7 @@ void SolGraf(TPZCompMesh *malha, std::ofstream &GraficoSol){
 		dsol[0].Redim(3,1);
 		
 		///Percorrer todos elementos 
-		for(int el=0; el < nelem; el++){
+		for(long el=0; el < nelem; el++){
 				TPZCompEl * Cel = malha->ElementVec()[el];
 				
 				if(!Cel) continue;
@@ -952,8 +952,8 @@ TPZGeoMesh * GeoMeshGrid( int h){
 		for ( int ref = 0; ref < h; ref++ )
 		{// h indica o numero de refinamentos
 				TPZVec<TPZGeoEl *> filhos;
-				int n = gmesh->NElements();
-				for ( int i = 0; i < n; i++ )
+				long n = gmesh->NElements();
+				for ( long i = 0; i < n; i++ )
 				{
 						TPZGeoEl * gel = gmesh->ElementVec() [i];
 						//if ( gel->Dimension() == 2 ) gel->Divide ( filhos );
